CU-9-Create-bottom-nav-bar. Created util package. Renamed NavigationPage to StubPage. Changed text style in BottomBar.
