+++ app/src/main/java/com/example/compassofukraine/MainActivity.kt
@@ -15,8 +15,8 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.navigation.compose.rememberNavController
 import com.example.compassofukraine.ui.BottomBar
-import com.example.compassofukraine.ui.BottomNavGraph
 import com.example.compassofukraine.ui.theme.CompassOfUkraineTheme
+import com.example.compassofukraine.util.BottomNavGraph
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
