+++ app/src/main/java/com/example/compassofukraine/ui/BottomBar.kt
@@ -21,6 +21,7 @@ import androidx.navigation.NavDestination.Companion.hierarchy
 import androidx.navigation.NavGraph.Companion.findStartDestination
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.currentBackStackEntryAsState
+import com.example.compassofukraine.util.BottomBarMenu
 
 @Composable
 fun BottomBar(navHostController: NavHostController) {
@@ -62,7 +63,7 @@ fun RowScope.AddItem(
         label = {
             Text(
                 text = screen.title,
-                style = MaterialTheme.typography.labelSmall,
+                style = MaterialTheme.typography.titleSmall,
                 maxLines = 1
             )
         },
